ARM64: review reworks
